pyproject.toml
[project]
name = "repl-nix-basicraganalyzer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.9.0",
    "langchain>=0.3.4",
    "openai>=1.52.2",
    "plotly>=5.24.1",
    "streamlit>=1.39.0",
]
